[Device] Change name to generic-device
