+++ include/CMakeLists.txt
@@ -14,6 +14,7 @@ SET(NEWHEADERS
   sot/core/debug.hh
   sot/core/derivator.hh  
   sot/core/device.hh
+  sot/core/generic-device.hh
   sot/core/event.hh
   sot/core/exception-abstract.hh
   sot/core/exception-dynamic.hh
