+++ unitTesting/tools/test_generic_device.cpp
@@ -18,7 +18,7 @@ using namespace std;
 
 #include <dynamic-graph/factory.h>
 #include <dynamic-graph/entity.h>
-#include <sot/core/device.hh>
+#include <sot/core/generic-device.hh>
 
 #include <sstream>
 #include <fstream>
@@ -183,7 +183,7 @@ void CreateYAMLFILE() {
   of.close();
 }
 
-int ReadYAMLFILE(dg::sot::Device &aDevice) {
+int ReadYAMLFILE(dg::sot::GenericDevice &aDevice) {
   // Reflect how the data are splitted in two yaml files in the sot
   // Comment and use the commented code to use the above yaml file
   std::ifstream yaml_file_controller("../../unitTesting/tools/sot_controller.yaml");
@@ -226,7 +226,7 @@ int main(int, char **) {
   robot_description = strStream.str();
 
   /// Test reading the URDF file.
-  dg::sot::Device aDevice(std::string("simple_humanoid"));
+  dg::sot::GenericDevice aDevice(std::string("simple_humanoid"));
   aDevice.setDebugMode(debug_mode);
   aDevice.setURDFModel(robot_description);
